Fixed file deletion error handling

+++ SwiftVideoGenerator/Classes/VideoGenerator.swift
@@ -243,11 +243,13 @@ public class VideoGenerator: NSObject {
       /// create a path to the video file
       completeMoviePath = URL(fileURLWithPath: documentsPath).appendingPathComponent("\(_fileName).m4v")
       
-      do {
-        /// delete an old duplicate file
-        try FileManager.default.removeItem(at: completeMoviePath!)
-      } catch {
-        failure(error)
+      if FileManager.default.fileExists(atPath: completeMoviePath!.absoluteString) {
+        do {
+          /// delete an old duplicate file
+          try FileManager.default.removeItem(at: completeMoviePath!)
+        } catch {
+          failure(error)
+        }
       }
     } else {
       failure(VideoGeneratorError(error: .kFailedToFetchDirectory))
